Add ambient hero components for academic, atlas, legal, FAQ, and whitepapers
